Introduce a distinction between total and unique tsize.
The total tsize of a file is the number of headers that
would be parsed if include guards were not in use. The
unique tsize is the number of files parsed taking include
guards into account.

A side effect of splitting <linux/kernel.h> and <linux/c.h>
is an increase of one in the unique tsize of many files. The
concept of total tsize is needed to accurately describe the
real reduction in complexity of these inclusion trees.

Add a 'visiting' stack to the child walker to prevent deep
recursion which was previously avoided by the unique tsize
semantics.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>