Correct the total tsize printed in child-depth-limited graphs.
The old total tsize calculation was done during the
output graph collection process.  This meant that if the
output graph was depth limited in the child direction,
the total tsize was not properly determined.

The new method uses the Graph class total_tsize() method
and prints correct total tsizes regardless of the child
depth limit in effect.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>