Improve edge snipping heuristics.
Modify the Analyser to return a sorted list of cutpoints
rather than just a count for each target.  This allows the
snipping logic to keep the link from the source(s) with the
largest unique tsizes and snip the rest.

This has the effect of pushing the yellow and red nodes up
against the blue backbone. The colored portion of the graph
should now be contiguous, with little tendrils of dark grey
"uninteresting" headers, plus a bunch of dashed back edges
to the yellow and red nodes situated "closer" to the root
node.

Also, reduce the acceptable indegree for the backbone to 2.
This also removes a lot of cross edges and helps clean
up the layout.  Reducing the indegree to 1 or increasing to
3 produces less appealing results, at least for linux/sched.h.
This might want to become tunable at some point, however.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>