Complete overhaul the node printing strategy.
Rather than creating ghost nodes for clipped children,
we simply make the parent of the ghost nodes into a
rectangle and list the names of all the children we
snipped from it. The snipped child list is now apparent
at a glance, rather than having to look for all the
ghost children.  This also reduces the number of nodes
enormously.

The snipped child nodes no longer have diamond or octagon
shapes.  we simply color them red or yellow, based on
their unique tsize. They are shaped as an ordinary
ellipse, unless they too have snipped children, in which
case they are box shaped as well.

In a nutshell, the snipped edges are "invisible edges
from box shaped nodes to red or yellow nodes"

Nodes are now described with the file name plus two or
three numbers:

  node name
  unique_tsize - total_tsize [ - num_snips ]

Nodes which have not been snipped have only the first two
numbers.  Nodes which have had incoming edges snipped
also have the third number which tells how many parent
links were removed.

This is the new meanings of node shape and color:

  shape:
       house - root node
         box - header file with snipped children
     ellipse - regular header file

  color:
      orange - root node
         red - large-tsize popular inclusion target
      yellow - small-tsize popular inclusion target
        blue - primary inclusion tree (backbone)
       green - regular header file

As before, more saturated means bigger tsize for red
and blue nodes.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>