Improve compactness of generated graph.
First change is to prefer light-weight edges instead of
heavy-weight edges in should_snip().  This has the effect
of pushing the yellow and red nodes as far away from the
blue backbone as possible.  It also decreases the edge
density near the root which improves the layout.

Second change is to snip all the children of the root
node if there are more than 3 present.  This eliminates
the bundle of edges converging on the root node and
greatly reduces the aspect ratio of complex graphs.

If the root node does get snipped, we introduce a series
of small orange circles to mark all the places that the
root node was originally connected to.

Unfortunately, the second change reintroduces the floating
subgraph problem....

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>