Use JPG instead of PostScript.
Postscript looks nicer but is much slower to render.
Graphviz doesn't seem to do DSC-compliant docs either,
so the files are a bit of a pain to work with.  Finally,
Windows doesn't have much in the way of good PostScript
tools, but JPG support should be fine on that platform.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>